+++ openff/nagl/config/training.py
@@ -1,3 +1,7 @@
+"""
+The `TrainingConfig` class
+"""
+
 from openff.nagl._base.base import ImmutableModel
 from openff.nagl.config.data import DataConfig
 from openff.nagl.config.model import ModelConfig
@@ -5,6 +9,9 @@ from openff.nagl.config.optimizer import OptimizerConfig
 from openff.nagl.utils._types import FromYamlMixin
 
 class TrainingConfig(ImmutableModel, FromYamlMixin):
+    """
+    Overall config class for training a model
+    """
     model: ModelConfig
     data: DataConfig
     optimizer: OptimizerConfig